Split DeferredTaskDefinitionIntegrationTest
